requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
psycopg2-binary
python-dotenv